claircore: bump to v0.2.0
This commit patches ClairCore to v0.2.0.

The AffectedManifests API is updated to pass back a pointer
AffectedManifest struct.

The DeleteUpdateOperation API is updated to return the number of deleted
UpdateOperations.

All occurences of TestLog are replaced with zlog.Test

Various mocks scattered around tests have been consolidated to their
appropriate service packages and removed.

+++ matcher/mock.go
@@ -12,8 +12,11 @@ import (
 var _ Service = (*Mock)(nil)
 
 // Mock implements a mock matcher service
+//
+// If a method is not provided to a constructed Mock a panic
+// will occur on call.
 type Mock struct {
-	DeleteUpdateOperations_ func(context.Context, ...uuid.UUID) error
+	DeleteUpdateOperations_ func(context.Context, ...uuid.UUID) (int64, error)
 	UpdateOperations_       func(context.Context, ...string) (map[string][]driver.UpdateOperation, error)
 	LatestUpdateOperation_  func(context.Context) (uuid.UUID, error)
 	LatestUpdateOperations_ func(context.Context) (map[string][]driver.UpdateOperation, error)
@@ -26,7 +29,10 @@ type Mock struct {
 }
 
 // DeleteUpdateOperations marks the provided refs as seen and processed.
-func (d *Mock) DeleteUpdateOperations(ctx context.Context, refs ...uuid.UUID) error {
+func (d *Mock) DeleteUpdateOperations(ctx context.Context, refs ...uuid.UUID) (int64, error) {
+	if d.DeleteUpdateOperations_ == nil {
+		panic("mock matcher: unexpected call to DeleteUpdateOperations")
+	}
 	return d.DeleteUpdateOperations_(ctx, refs...)
 }
 
@@ -34,25 +40,40 @@ func (d *Mock) DeleteUpdateOperations(ctx context.Context, refs ...uuid.UUID) er
 //
 // "Prev" can be `uuid.Nil` to indicate "earliest known ref."
 func (d *Mock) UpdateDiff(ctx context.Context, prev uuid.UUID, cur uuid.UUID) (*driver.UpdateDiff, error) {
+	if d.UpdateDiff_ == nil {
+		panic("mock matcher: unexpected call to UpdateDiff")
+	}
 	return d.UpdateDiff_(ctx, prev, cur)
 }
 
 // UpdateOperations returns all the known UpdateOperations per updater.
 func (d *Mock) UpdateOperations(ctx context.Context, updaters ...string) (map[string][]driver.UpdateOperation, error) {
+	if d.UpdateOperations_ == nil {
+		panic("mock matcher: unexpected call to UpdateOperations")
+	}
 	return d.UpdateOperations_(ctx, updaters...)
 }
 
 // LatestUpdateOperations returns the most recent UpdateOperation per updater.
 func (d *Mock) LatestUpdateOperations(ctx context.Context) (map[string][]driver.UpdateOperation, error) {
+	if d.LatestUpdateOperations_ == nil {
+		panic("mock matcher: unexpected call to LatestUpdateOperations")
+	}
 	return d.LatestUpdateOperations_(ctx)
 }
 
 // LatestUpdateOperation returns a ref for the most recent update operation
 // across all updaters.
 func (d *Mock) LatestUpdateOperation(ctx context.Context) (uuid.UUID, error) {
+	if d.LatestUpdateOperation_ == nil {
+		panic("mock matcher: unexpected call to LatestUpdateOperation")
+	}
 	return d.LatestUpdateOperation_(ctx)
 }
 
-func (s *Mock) Scan(ctx context.Context, ir *claircore.IndexReport) (*claircore.VulnerabilityReport, error) {
-	return s.Scan_(ctx, ir)
+func (d *Mock) Scan(ctx context.Context, ir *claircore.IndexReport) (*claircore.VulnerabilityReport, error) {
+	if d.Scan_ == nil {
+		panic("mock matcher: unexpected call to Scan")
+	}
+	return d.Scan_(ctx, ir)
 }
